【ISSUE #278】Define and Implement the RmqConnector and RmqSourceTask. (#381)
* init commit

* [Fix]fix some bugs in connectors and fastjson.

* [Update]update the version of replicator

* [Update]remove the repetitive properties

* [Update]Add the license at the head of TaskTopicInfo class and TaskConfigEnum class

* [Update]update the project name from  to

* [Update]update the language of README.md

* [Update]change the title of README.md

* [Update]change the package name from  to

* [Update]update the  starting position

* [Update]update the  starting position

* [Fix]fix the bug of  instanceName setting
