+++ rocketmq-replicator/pom.xml
@@ -22,8 +22,8 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.apache.rocketmq</groupId>
-    <artifactId>rocketmq-connector</artifactId>
-    <version>0.0.1-SNAPSHOT</version>
+    <artifactId>rocketmq-replicator</artifactId>
+    <version>0.1.0-SNAPSHOT</version>
 
     <properties>
         <rocketmq.version>4.4.0</rocketmq.version>
