+++ rocketmq-replicator/src/main/java/org/apache/rocketmq/replicator/strategy/DivideStrategyEnum.java
@@ -14,29 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.rocketmq.connector.config;
+package org.apache.rocketmq.replicator.strategy;
 
-public enum TaskConfigEnum {
+public enum DivideStrategyEnum {
 
-    TASK_ID("taskId"),
-    TASK_SOURCE_GROUP("sourceGroup"),
-    TASK_SOURCE_ROCKETMQ("sourceRocketmq"),
-    TASK_SOURCE_TOPIC("sourceTopic"),
-    TASK_STORE_ROCKETMQ("storeTopic"),
-    TASK_DATA_TYPE("dataType"),
-    TASK_BROKER_NAME("brokerName"),
-    TASK_QUEUE_ID("queueId"),
-    TASK_NEXT_POSITION("nextPosition"),
-    TASK_TOPIC_INFO("taskTopicList"),
-    TASK_SOURCE_RECORD_CONVERTER("source-record-converter");
-
-    private String key;
-
-    TaskConfigEnum(String key) {
-        this.key = key;
-    }
-
-    public String getKey() {
-        return key;
-    }
+    BY_TOPIC,
+    BY_QUEUE
 }
